Link containers via explicit env var host refs
Docker-compose and Kubernetes use very different ways of linking containers.

To be compatible with both, explicitly pass the hostname to refer to in an
environment variable - specified in our orchestration file (docker-compose or
Kubernetes pod definition), not the end-deployer configuration file.

The variables are DJNRO_HOST and DB_HOST and should either point to the link
name (djnro/postgres) for docker-compose or localhost for Kubernetes.