Kuberentes: Postgres: override PGDATA into a subdir
Postgres does not like using a non-empty /var/lib/postgresql/data

Kubernetes/GCE create a persistent disk with a /lost+found directory.

The documentation at https://hub.docker.com/r/library/postgres/
recommends in that case to override PGDATA to point to a subdirectory, e.g.
/var/lib/postgresql/data/pgdata).

Implemented exactly as per this suggestion as a variable in the pod definition.